+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Version counting is based on semantic versioning (Major.Feature.Patch)
 
 ### YACReaderLibraryServer
 * Don't crash at startup if the server can't bind to any address.
+* Redirect `/` to `/webui`.
 
 ### All GUI Apps
 * Better system theme detection.